Extract HPPManagerError to separate file
