+++ Pod/Classes/RealexComponent/HPPManager.swift
@@ -409,22 +409,3 @@ private struct HPPHeader {
         static let all = "*/*"
     }
 }
-
-@objcMembers public class HPPManagerError: NSError {
-
-    public static func missingProducerURL() -> NSError {
-        NSError(
-            domain: "com.realex.payments",
-            code: 9000,
-            userInfo: [NSLocalizedDescriptionKey : "HPPRequestProducerURL can't be blank"]
-        )
-    }
-
-    public static func typeMismatch() -> NSError {
-        NSError(
-            domain: "com.realex.payments",
-            code: 9001,
-            userInfo: [NSLocalizedDescriptionKey : "decodedResponse should be a [String: String]"]
-        )
-    }
-}
